+++ qunit/tests_alternations.js
@@ -121,28 +121,16 @@ define([
 		assert.equal(testmask.inputmask._valueGet(), "WZ", "Result " + testmask.inputmask._valueGet());
 	});
 
-	qunit.test("alternations (9{1,3})|(S|NE|W)-9{1,3}", function (assert) {
+	qunit.test("alternations (9{1,3}|SE|NE|SW|NW)-9{1,3} - yesman85", function (assert) {
 		var $fixture = $("#qunit-fixture");
 		$fixture.append('<input type="text" id="testmask" />');
 		var testmask = document.getElementById("testmask");
 
-		Inputmask("(9{1,3})|(S|NE|W)-9{1,3}").mask(testmask);
-
-		$("#testmask").Type("NE123");
-
-		assert.equal(testmask.inputmask._valueGet(), "NE-123", "Result " + testmask.inputmask._valueGet());
-	});
-
-	qunit.test("alternations (9{1,3})|((S|N)(E|W))-9{1,3} - yesman85", function (assert) {
-		var $fixture = $("#qunit-fixture");
-		$fixture.append('<input type="text" id="testmask" />');
-		var testmask = document.getElementById("testmask");
-
-		Inputmask("(9{1,3})|((S|N)(E|W))-9{1,3}").mask(testmask);
+		Inputmask("(9{1,3}|SE|NE|SW|NW)-9{1,3}").mask(testmask);
 
 		$("#testmask").Type("(NE123");
 
-		assert.equal(testmask.inputmask._valueGet(), "(N)(E)-123", "Result " + testmask.inputmask._valueGet());
+		assert.equal(testmask.inputmask._valueGet(), "NE-123", "Result " + testmask.inputmask._valueGet());
 	});
 
 	qunit.test("((S))", function (assert) {
@@ -305,12 +293,13 @@ define([
 		$("#testmask").Type("23");
 		assert.equal(testmask.inputmask._valueGet(), "23", "Result " + testmask.inputmask._valueGet());
 	});
-	qunit.test("(9)|(09)|(19)|(2f) - type 24 - ivaninDarpatov", function (assert) {
+
+	qunit.test("(9|09|19|2f) - type 24 - ivaninDarpatov", function (assert) {
 		var $fixture = $("#qunit-fixture");
 		$fixture.append('<input type="text" id="testmask" />');
 		var testmask = document.getElementById("testmask");
 
-		Inputmask("(9)|(09)|(19)|(2f)", {
+		Inputmask("(9|09|19|2f)", {
 			definitions: {
 				"f": {validator: "[0-3]"}
 			}